# Implement a sorting queue using a loser tree to improve merge sort performance