--- a/src/Core/SortCursor.cpp
+++ b/src/Core/SortCursor.cpp
@@ -48,8 +48,8 @@ void SortCursorImpl::reset(const Columns & columns, const Block & block, UInt64 
     permutation = perm;
 }
 
-template <bool use_loser_tree>
-DataTypes SortQueueVariants<use_loser_tree>::extractSortDescriptionTypesFromHeader(const Block & header, const SortDescription & sort_description)
+template <QueueImplType queue_type>
+DataTypes SortQueueVariants<queue_type>::extractSortDescriptionTypesFromHeader(const Block & header, const SortDescription & sort_description)
 {
     size_t sort_description_size = sort_description.size();
     DataTypes data_types(sort_description_size);
@@ -62,7 +62,7 @@ DataTypes SortQueueVariants<use_loser_tree>::extractSortDescriptionTypesFromHead
 
     return data_types;
 }
-template class SortQueueVariants<false>;
-template class SortQueueVariants<true>;
+template class SortQueueVariants<QueueImplType::Default>;
+template class SortQueueVariants<QueueImplType::LoserTree>;
 
 }