--- a/src/Processors/Merges/Algorithms/MergingSortedAlgorithm.cpp
+++ b/src/Processors/Merges/Algorithms/MergingSortedAlgorithm.cpp
@@ -53,7 +53,7 @@ MergingSortedAlgorithm::MergingSortedAlgorithm(
         sort_description_types.emplace_back(header->getByName(column_description.column_name).type);
     }
 
-    queue_variants = SortQueueVariants<QueueImplType::LoserTree>(sort_description_types, description);
+    queue_variants = SortQueueVariants<>(sort_description_types, description);
     if (queue_variants.variantSupportJITCompilation())
         compileSortDescriptionIfNeeded(description, sort_description_types, true /*increase_compile_attempts*/);
 }