--- a/src/Processors/Merges/Algorithms/MergingSortedAlgorithm.h
+++ b/src/Processors/Merges/Algorithms/MergingSortedAlgorithm.h
@@ -62,7 +62,7 @@ private:
 
     SortCursorImpls cursors;
 
-    SortQueueVariants queue_variants;
+    SortQueueVariants<> queue_variants;
 
     template <typename TSortingQueue>
     Status mergeImpl(TSortingQueue & queue);