--- a/src/Processors/Transforms/SortingTransform.cpp
+++ b/src/Processors/Transforms/SortingTransform.cpp
@@ -195,7 +195,7 @@ SortingTransform::SortingTransform(
 
     description.swap(description_without_constants);
 
-    if (SortQueueVariants<true>(sort_description_types, description).variantSupportJITCompilation())
+    if (SortQueueVariants<QueueImplType::LoserTree>(sort_description_types, description).variantSupportJITCompilation())
         compileSortDescriptionIfNeeded(description, sort_description_types, increase_sort_description_compile_attempts /*increase_compile_attempts*/);
 }