--- a/src/Processors/Transforms/SortingTransform.h
+++ b/src/Processors/Transforms/SortingTransform.h
@@ -17,8 +17,6 @@ class MergeSorter
 public:
     MergeSorter(SharedHeader header, Chunks chunks_, SortDescription & description_, size_t max_merged_block_size_, UInt64 limit_);
 
-    ~MergeSorter();
-
     Chunk read();
 
 private: