@lgbo-ustc 2025-09-10
Using a loser tree can cut the comparison operations almost in half.  It ​​can improve​​ performance when the sort keys ​​have​​ high cardinality.

- before
```
0 rows in set. Elapsed: 6.599 sec. Processed 10.00 million rows, 377.25 MB (1.52 million rows/s., 57.17 MB/s.)
```

- after
```
0 rows in set. Elapsed: 5.535 sec. Processed 10.00 million rows, 377.25 MB (1.81 million rows/s., 68.16 MB/s.)
```